requirements.txt
APScheduler==3.7.0
autopep8==1.5.5
bottle==0.12.19
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.0
idna==3.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyDrive==1.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.0.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
tweepy==3.10.0
twitter==1.18.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
